+++ inc/classes/class-load-scripts.php
@@ -110,16 +110,33 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 		private $local = [];
 
 		/**
+		 * Inline scripts
+		 *
+		 * @var array $inline
+		 */
+		private $inline = [];
+
+		/**
+		 * Script attributes - defer, async, integrity
+		 *
+		 * @var array $attr
+		 */
+		private $attr = [];
+
+		/**
 		 * Class constructor
 		 * - set up hooks
 		 */
 		public function __construct() {
 
-			// Load adminscripts
+			// Load admin scripts
 			add_action( 'admin_enqueue_scripts', 	[ $this, 'load_admin_scripts' ] ); 
 
 			// Login page scripts
-			add_action( 'login_enqueue_scripts', 	[ $this, 'load_login_scripts' ], 1 );			
+			add_action( 'login_enqueue_scripts', 	[ $this, 'load_login_scripts' ], 1 );
+
+			// Add attributes to scripts if there			
+			add_filter( 'script_loader_tag', 		[ $this, 'add_scripts_attr' ], 10, 3 );
 
 			// Front end only
 			if ( is_admin() ) { return; }
@@ -131,7 +148,7 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			add_action( 'wp_enqueue_scripts', 		[ $this, 'load_child_scripts' ], 25 ); 
 
 			// Dequeueue scripts
-			add_action( 'wp_enqueue_scripts', 		[ $this, 'undo_scripts' ], 99 ); 
+			add_action( 'wp_enqueue_scripts', 		[ $this, 'undo_scripts' ], 20 ); 
 
 			// Conditional header scripts
 			add_action( 'wp_enqueue_scripts', 		[ $this, 'conditional_scripts' ] ); 
@@ -189,8 +206,14 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			// Login scripts: [ 'label' => [ 'src', (array)deps, 'ver' ] ... ]
 			$this->login = $this->set_key( $scripts, 'login' );
 
-			// Localize scripts: [ 'label' => [ 'name' => name, trans => function/src ] ]
+			// Localize scripts: [ 'handle' => [ 'name' => name, trans => function/src ] ]
 			$this->local = $this->set_key( $scripts, 'local' );
+
+			// Inline scripts: [ 'handle' => [ 'src' => function/src ] ]
+			$this->inline = $this->set_key( $scripts, 'inline' );
+
+			// Localize scripts: [ 'label' => [ 'handle' ] ]
+			$this->attr = $this->set_key( $scripts, 'attr' );
 		}
 
 		/**
@@ -224,8 +247,9 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 				// Test hook dependency
 				if ( !empty( $v[0] ) && $hook != $v[0] ) { continue; }
 
-				// Register and enqueue scripts in footer
-				wp_register_script( $k, $v[1], $v[2], $v[3], true ); 
+				// Register and enqueue scripts in footer by default
+				$locale = ( isset( $v[4] ) && $v[4] === false ) ? false : true;			   	
+				wp_register_script( $k, $v[1], $v[2], $v[3], $locale ); 
 				wp_enqueue_script( $k );
 			}
 		}
@@ -247,9 +271,10 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 
 			// Set up core localization
 			$trans = [ 
-				'home_url' => home_url(), 
-				'ajax_url' => admin_url( 'admin-ajax.php' ),
-				'rest_url' => rest_url( '/' ) 
+				'home_url' 	=> home_url(), 
+				'ajax_url' 	=> admin_url( 'admin-ajax.php' ),
+				'rest_url' 	=> rest_url( '/' ),
+			    'is_admin'	=> is_admin()
 			];
 			wp_localize_script( 'ipress', 'theme', $trans ); 
 
@@ -269,6 +294,9 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			foreach ( $this->external as $k=>$v ) { 
 				$locale = ( isset( $v[3] ) && $v[3] === true ) ? true : false;
 				wp_register_script( $k, $v[0], $v[1], $v[2], $locale ); 
+				if ( array_key_exists( $k, $this->inline ) ) {
+					$this->inline( $k );
+				}
 				wp_enqueue_script( $k );
 			}
 
@@ -292,17 +320,19 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 
 			// Register & enqueue plugin scripts
 			foreach ( $this->plugins as $k=>$v ) { 
-				wp_register_script( $k, $v[0], $v[1], $v[2], true ); 
+				$locale = ( isset( $v[3] ) && $v[3] === false ) ? false : true;
+				wp_register_script( $k, $v[0], $v[1], $v[2], $locale ); 
 				if ( array_key_exists( $k, $this->local ) ) {
 					$this->localize( $k );
 				}
 				wp_enqueue_script( $k );
 			}
 
-			// Page templates in footer head
+			// Page templates in footer / header
 			foreach ( $this->page as $k=>$v ) {
 				if ( is_page_template( $v[0] ) ) {
-					wp_register_script( $k, $v[1], $v[2], $v[3], true ); 
+					$locale = ( isset( $v[4] ) && $v[4] === false ) ? false : true;			   	
+					wp_register_script( $k, $v[1], $v[2], $v[3], $locale ); 
 					if ( array_key_exists( $k, $this->local ) ) {	
 						$this->localize( $k );
 					} 
@@ -324,12 +354,13 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 				}
 			}
 
-			// Front page scripts
+			// Front page scripts in footer by default
 			foreach ( $this->front as $k=>$v ) {
 				switch( $v[0] ) {
 					case 'front' :
-						if ( is_front_page() ) { 
-							wp_register_script( $k, $v[1], $v[2], $v[3], true ); 
+						if ( is_front_page() ) {
+							$locale = ( isset( $v[4] ) && $v[4] === false ) ? false : true;			   	
+							wp_register_script( $k, $v[1], $v[2], $v[3], $locale ); 
 							if ( array_key_exists( $k, $this->local ) ) {
 								$this->localize( $k );
 							}
@@ -338,7 +369,8 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 						break;
 					case 'home' :
 						if ( is_home() ) { 
-							wp_register_script( $k, $v[1], $v[2], $v[3], true ); 
+							$locale = ( isset( $v[4] ) && $v[4] === false ) ? false : true;			   	
+							wp_register_script( $k, $v[1], $v[2], $v[3], $locale ); 
 							if ( array_key_exists( $k, $this->local ) ) {
 								$this->localize( $k );
 							}
@@ -348,7 +380,8 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 					case 'front-home' :
 					case 'home-front' :
 						if ( is_home() && is_front_page() ) { 
-							wp_register_script( $k, $v[1], $v[2], $v[3], true ); 
+							$locale = ( isset( $v[4] ) && $v[4] === false ) ? false : true;			   	
+							wp_register_script( $k, $v[1], $v[2], $v[3], $locale ); 
 							if ( array_key_exists( $k, $this->local ) ) {
 								$this->localize( $k );
 							}
@@ -392,6 +425,23 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 		}
 
 		/**
+		 * Add inline scripts
+		 *
+		 * @param string $key
+		 * @return void
+		 */
+		private function inline( $key ) {
+
+			// Get local key
+			$h = $this->inline[$key]; 
+
+			// Validate & Inject Inline Script
+			if ( isset( $h['src'] ) ) { 
+				wp_add_inline_script( $key, $h['src'] ); 
+			}
+		}
+
+		/**
 		 * Dequeue scripts 
 		 */
 		public function undo_scripts() { 
@@ -431,8 +481,8 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			if ( ! $support_old_ie ) { return; }
 
 			// Enqueue scripts
-			wp_enqueue_script( 'html5-shiv', 'https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js', [], NULL );
-			wp_enqueue_script( 'respond-min', 'https://oss.maxcdn.com/libs/respond.js/1.3.0/respond.min.js', [], NULL );
+			wp_enqueue_script( 'html5-shiv', 'https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js', [], null, false );
+			wp_enqueue_script( 'respond-min', 'https://oss.maxcdn.com/libs/respond.js/1.3.0/respond.min.js', [], null, false );
 
 			// Add to global scripts list
 			$wp_scripts->add_data( 'html5-shiv', 'conditional', 'lt IE 9' );
@@ -449,7 +499,7 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 		 */
 		public function header_scripts() {
 
-			// Set?
+			// Get content?
 			$scripts = apply_filters( 'ipress_header_scripts', get_theme_mod( 'ipress_header_scripts', '' ) );
 			if ( empty( $scripts ) ) { return; }
 			
@@ -463,7 +513,7 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 		 */
 		public function footer_scripts() {
 
-			// Set?
+			// Get content?
 			$scripts = apply_filters( 'ipress_footer_scripts', get_theme_mod( 'ipress_footer_scripts', '' ) );
 			if ( empty( $scripts ) ) { return; }
 
@@ -510,10 +560,62 @@ if ( ! class_exists( 'IPR_Load_Scripts' ) ) :
 			// Register & enqueue admin scripts
 			foreach ( $this->login as $k=>$v ) { 
 				
-				// Register and enqueue script
-				wp_register_script( $k, $v[0], $v[1], $v[2], true ); 
+				// Register and enqueue script in footer by default
+				$locale = ( isset( $v[3] ) && $v[3] === false ) ? false : true;			   	
+				wp_register_script( $k, $v[0], $v[1], $v[2], $locale ); 
 				wp_enqueue_script( $k );
 			}
+		}
+
+		//----------------------------------------------
+		//	Script attributes - defer, async, integrity
+		//----------------------------------------------
+
+		/**
+		 * Tag on script attributes to matching handles
+		 *
+		 * @param	string	$tag
+		 * @param	string	$handle
+		 * @param	string	$src
+		 * @return	string
+		 */
+		public function add_scripts_attr( $tag, $handle, $src ) {
+
+			// Nothing set?
+			if ( empty( $this->attr ) ) { return $tag; }
+
+			// Sort attr
+			$defer 		= ( isset( $this->attr['defer'] ) && ! empty( $this->attr['defer'] ) ) ? $this->attr['defer'] : [];
+			$async 		= ( isset( $this->attr['async'] ) && ! empty( $this->attr['async'] ) ) ? $this->attr['async'] : [];
+			$integrity 	= ( isset( $this->attr['integrity'] ) && ! empty( $this->attr['integrity'] ) ) ? $this->attr['integrity'] : [];
+
+			// Initialise attr
+			$attr = [];
+
+			// Ok, do defer
+			if ( ! empty( $defer ) && in_array( $handle, $defer ) ) {
+				$attr[] = 'defer="defer"';
+			}
+
+			// Ok, do async
+			if ( ! empty( $async ) && in_array( $handle, $async ) ) {
+				$attr[] = 'async="async"';
+			}
+
+			// Ok, do integrity
+			if ( ! empty( $integrity ) ) {
+				foreach ( $integrity as $k=>$v ) {
+					foreach ( $v as $h=>$a ) {
+						if ( $handle === $h ) {
+							$attr[] = sprintf( 'integrity="%s" crossorigin="anonymous"', $a );
+							break;
+						}
+					}
+				}
+			}
+
+			// Return script tag
+			return ( empty( $attr ) ) ? $tag : sprintf( '<script src="%s" %s></script>' . "\n", $src, join( ' ', $attr ) );
 		}
 	}