+++ search.php
@@ -18,11 +18,9 @@
 
 <?php get_header(); ?>
 
-<div id="primary" class="content-area">
-
 <?php do_action( 'ipress_before_main_content' ); ?>
 
-	<main id="main" class="site-main" role="main">
+	<main id="main" class="site-main">
 
 	<?php do_action( 'ipress_search_before' ); ?>
 
@@ -30,9 +28,9 @@
 
 		<header class="page-header">
 			<h1 class="page-title search-title">
-			<?php 
+			<?= 
 				/* translators: %s: search query. */
-				echo sprintf( esc_html__( 'Search: %s Results for <span>%s</span>', 'ipress' ), $wp_query->found_posts, get_search_query() ); ?>
+				sprintf( esc_html__( 'Search: %s Results for <span>%s</span>', 'ipress' ), $wp_query->found_posts, get_search_query() ); ?>
 			</h1>
 		</header><!-- .page-header -->
 
@@ -46,12 +44,10 @@
 
 	<?php do_action( 'ipress_search_after' ); ?>
 
-	</main><!-- #main -->
+	</main><!-- #main / .site-main -->
 
 <?php do_action( 'ipress_after_main_content' ); ?>
 
-</div><!-- #primary -->
+<?php do_action( 'ipress_sidebar' ); ?>
 
-<?php
-do_action( 'ipress_sidebar' );
-get_footer();
+<?php get_footer();