+++ templates/post/content-excerpt.php
@@ -12,9 +12,10 @@
  * @license     GPL-2.0+
  */
 ?>
-<?php do_action( 'ipress_loop_post_before' ); ?>
 
-<article id="post-<?php the_ID(); ?>" <?php post_class( 'entry' ); ?>>
+<?php do_action( 'ipress_loop_before' ); ?>
+
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?> itemscope itemtype="https://schema.org/CreativeWork">
 
 <?php
 	/**
@@ -24,8 +25,8 @@
 	 * @hooked ipress_loop_excerpt 	- 20
 	 * @hooked ipress_loop_footer 	- 30
 	 */
-	do_action( 'ipress_loop_post_excerpt' );
-?>
+	do_action( 'ipress_loop_excerpt' ); ?>
+
 </article><!-- #post-<?php the_ID(); ?> -->
 
-<?php do_action( 'ipress_loop_post_after' );
+<?php do_action( 'ipress_loop_after' );