+++ templates/post/content-image.php
@@ -12,9 +12,11 @@
  * @license     GPL-2.0+
  */
 ?>
-<?php do_action( 'ipress_loop_post_before' ); ?>
 
-<article id="post-<?php the_ID(); ?>" <?php post_class( 'entry' ); ?>>
+<?php do_action( 'ipress_article_before' ); ?>
+
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?> itemscope itemtype="https://schema.org/CreativeWork">
+
 <?php
 	/**
 	 * Functions hooked in to ipress_loop_post action.
@@ -24,8 +26,8 @@
 	 * @hooked ipress_loop_gallery 		- 20
 	 * @hooked ipress_loop_footer 		- 30
 	 */
-	do_action( 'ipress_loop_post_image' );
-?>
+	do_action( 'ipress_loop_image' ); ?>
+
 </article><!-- #post-<?php the_ID(); ?> -->
 
-<?php do_action( 'ipress_loop_post_after' );
+<?php do_action( 'ipress_article_after' );