+++ templates/search/content.php
@@ -12,9 +12,10 @@
  * @license     GPL-2.0+
  */
 ?>
-<?php do_action( 'ipress_loop_post_before' ); ?>
 
-<article id="post-<?php the_ID(); ?>" <?php post_class( 'entry' ); ?>>
+<?php do_action( 'ipress_article_before' ); ?>
+
+<article id="post-<?php the_ID(); ?>" <?php post_class(); ?> itemscope itemtype="https://schema.org/CreativeWork">
 
 	<?php
 	/**
@@ -25,8 +26,8 @@
 	 * @hooked ipress_loop_excerpt      - 30
 	 * @hooked ipress_loop_footer       - 40
 	 */
-    do_action( 'ipress_loop_search' ); ?>
+    do_action( 'ipress_search' ); ?>
 
 </article><!-- #post-<?php the_ID(); ?> -->
 
-<?php do_action( 'ipress_loop_post_after' );
+<?php do_action( 'ipress_article_after' );