+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 * (07 August 2018). 1.0.0 Initial Parent Theme Fork 
 * (09 August 2018). 1.0.1 Refactor script & style dequeue
 * (13 August 2018). 1.0.2 Refactor config file
+* (26 September 2018). 1.0.3 Extend Scripts & Styles importers